requirements.txt
keyboard==0.13.5
Pillow==10.1.0
Pillow==10.2.0
PyAutoGUI==0.9.54
pystray==0.19.5
pytesseract==0.3.10
